Fix tests and add more resource monitor tests
