+++ test/qlever/resource_usage/test_usage_plot.py
@@ -13,7 +13,6 @@ from qlever.resource_usage.usage_plot import (  # noqa: E402
     read_usage_tsv,
     render_usage_plot,
 )
-from qlever.util import parse_git_hash  # noqa: E402
 
 
 def write_log(tmp_path, lines):
@@ -38,30 +37,6 @@ def write_log(tmp_path, lines):
 )
 def test_pick_time_unit(seconds, expected):
     assert pick_time_unit(seconds) == expected
-
-
-@pytest.mark.parametrize(
-    "first_line,expected",
-    [
-        ("qlever-server, git hash 1a2b3c4, compiled", "1a2b3c4"),
-        ("no hash on this line", None),
-    ],
-)
-def test_parse_git_hash_reads_first_line_only(first_line, expected, tmp_path):
-    path = tmp_path / "index-log.txt"
-    # Second line also carries a hash; only the first line should count.
-    path.write_text(first_line + "\nsomething git hash deadbeef here\n")
-    assert parse_git_hash(path) == expected
-
-
-def test_parse_git_hash_missing_file(tmp_path):
-    assert parse_git_hash(tmp_path / "nope.txt") is None
-
-
-def test_parse_git_hash_empty_file(tmp_path):
-    path = tmp_path / "empty.txt"
-    path.write_text("")
-    assert parse_git_hash(path) is None
 
 
 @pytest.mark.parametrize(
